Remove broken bucket as hostname detection code.
The code that detects if a bucket can be used as a hostname is
non optimal for SSL.  All buckets can be accessed via an endpoint
and a bucketname, or via a hostname.

When using SSL accessing buckets that contain periods in their name
the wildcard validation does not match causing SSL errors.

Also accessing buckets via hostname causes connection caching to not
work as well, since there are different connections for every S3 bucket
since they are connected to different hostnames.

This change does require the correct S3 endpoint to be specified for
a bucket if it is the US standard region.

Endpoints listed here:

http://docs.aws.amazon.com/general/latest/gr/rande.html#s3_region